+++ bootique/src/main/java/io/bootique/BQRuntimeListener.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed to ObjectStyle LLC under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ObjectStyle LLC licenses
+ * this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.bootique;
+
+/**
+ * A callback invoked after BQRuntime is created, but before any commands are run. Used primarily by the test tools
+ * and such to manage lifecycles external to Bootique. Most regular apps do not require this callback.
+ *
+ * @since 3.0.M1
+ */
+@FunctionalInterface
+public interface BQRuntimeListener {
+
+    void onRuntimeCreated(BQRuntime runtime);
+}