+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #315 Standard Modules should not have multiple constructors
 * #316 JUnit 4 support deprecation
 * #317 Java 11 as a minimal requirement for Bootique
+* #322 "BQRuntimeListener" to receive notifications on runtime startup
 
 ## 2.0